:recycle: Refactor Home, Not found and Gallery
