+++ src/notFound/pages/section/index.css
