+++ src/shared/assets/images/background/background.png
