+++ src/waiting/components/ImageGallery/Image.js
