+++ src/waiting/components/TextGallery/Text.js
