+++ src/waiting/pages/Gallery/WaitingTimeGallery.js
@@ -1,28 +1,27 @@
 import React, { useState } from 'react'
-import Image from 'Components/WaitingTime/Gallery/Image'
-import Text from 'Components/WaitingTime/Gallery/Text'
-import Title from 'assets/images/WaitingTime/gallery/Title.png'
-import Image1 from 'assets/images/WaitingTime/thumbnail/1.png'
-import Image2 from 'assets/images/WaitingTime/thumbnail/2.png'
-import Image3 from 'assets/images/WaitingTime/thumbnail/3.png'
-import Image4 from 'assets/images/WaitingTime/thumbnail/4.png'
-import Image5 from 'assets/images/WaitingTime/thumbnail/5.png'
-import Image6 from 'assets/images/WaitingTime/thumbnail/6.png'
-import Image7 from 'assets/images/WaitingTime/thumbnail/7.png'
-import Image8 from 'assets/images/WaitingTime/thumbnail/8.png'
-import Image9 from 'assets/images/WaitingTime/thumbnail/9.png'
-import Image10 from 'assets/images/WaitingTime/thumbnail/10.png'
-import ImageGallery1 from 'assets/images/WaitingTime/gallery/Image1.png'
-import ImageGallery2 from 'assets/images/WaitingTime/gallery/Image2.png'
-import ImageGallery3 from 'assets/images/WaitingTime/gallery/Image3.png'
-import ImageGallery4 from 'assets/images/WaitingTime/gallery/Image4.png'
-import ImageGallery5 from 'assets/images/WaitingTime/gallery/Image5.png'
-import ImageGallery6 from 'assets/images/WaitingTime/gallery/Image6.png'
-import ImageGallery7 from 'assets/images/WaitingTime/gallery/Image7.png'
-import ImageGallery8 from 'assets/images/WaitingTime/gallery/Image8.png'
-import ImageGallery9 from 'assets/images/WaitingTime/gallery/Image9.png'
-import ImageGallery10 from 'assets/images/WaitingTime/gallery/Image10.png'
-import Background from 'assets/images/WaitingTime/gallery/background.png'
+import Image from 'waiting/components/ImageGallery/Image'
+import Text from 'waiting/components/TextGallery/Text'
+import Title from 'waiting/assets/images/gallery/Title.png'
+import Image1 from 'waiting/assets/images/gallery/1.png'
+import Image2 from 'waiting/assets/images/gallery/2.png'
+import Image3 from 'waiting/assets/images/gallery/3.png'
+import Image4 from 'waiting/assets/images/gallery/4.png'
+import Image5 from 'waiting/assets/images/gallery/5.png'
+import Image6 from 'waiting/assets/images/gallery/6.png'
+import Image7 from 'waiting/assets/images/gallery/7.png'
+import Image8 from 'waiting/assets/images/gallery/8.png'
+import Image9 from 'waiting/assets/images/gallery/9.png'
+import Image10 from 'waiting/assets/images/gallery/10.png'
+import ImageGallery1 from 'waiting/assets/images/gallery/Image1.png'
+import ImageGallery2 from 'waiting/assets/images/gallery/Image2.png'
+import ImageGallery3 from 'waiting/assets/images/gallery/Image3.png'
+import ImageGallery4 from 'waiting/assets/images/gallery/Image4.png'
+import ImageGallery5 from 'waiting/assets/images/gallery/Image5.png'
+import ImageGallery6 from 'waiting/assets/images/gallery/Image6.png'
+import ImageGallery7 from 'waiting/assets/images/gallery/Image7.png'
+import ImageGallery8 from 'waiting/assets/images/gallery/Image8.png'
+import ImageGallery9 from 'waiting/assets/images/gallery/Image9.png'
+import ImageGallery10 from 'waiting/assets/images/gallery/Image10.png'
 import './WaitingTimeGallery.css'
 
 import Modal from 'shared/components/Modal/Modal'
